Rename identification file
